% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_forestplot.R
\name{ng_colour}
\alias{ng_colour}
\title{Nightingale's colours}
\usage{
ng_colour(...)
}
\arguments{
\item{...}{Character names of Nightingale's colours.}
}
\description{
\code{ng_colour()} returns Nightingale's colours' hex codes.
\code{display_ng_colours()} displays the avaliable colours, with their names
and hex codes, in a vertical layout.
}
\author{
Ilari Scheinin
}
