% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tabler.R
\name{tablerDashBody}
\alias{tablerDashBody}
\title{overwrite default tablerDashBody}
\usage{
tablerDashBody(...)
}
\arguments{
\item{...}{element to put inside Dash body}
}
\description{
overwrite default tablerDashBody
}
