% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tabler.R
\name{tablerDashFooter}
\alias{tablerDashFooter}
\title{Create a Boostrap 4 dashboard footer. Modification version of}
\usage{
tablerDashFooter(..., copyrights = NULL)
}
\arguments{
\item{...}{Left text.}

\item{copyrights}{Copyrights, if any.}
}
\description{
Build an adminLTE3 dashboard footer
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
