% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tabler.R
\name{tablerDashNavCondensed}
\alias{tablerDashNavCondensed}
\title{A condensed version of tablerDashNav.}
\usage{
tablerDashNavCondensed(id, ..., src = NULL, navMenu = NULL)
}
\arguments{
\item{id}{Navbar id}

\item{...}{Navbar content (will be placed on the right side of header)}

\item{src}{logo image source (displayed on the left side)}

\item{navMenu}{Slot for tablerNavMenu}
}
\description{
One header only and not a navbar below the header
}
