% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tabler.R
\name{tablerNavMenuDrop}
\alias{tablerNavMenuDrop}
\title{Build a tabler dashboard Dropdown menu to be placed inside tablerNavMenu()}
\usage{
tablerNavMenuDrop(..., text = NULL, icon = NULL)
}
\arguments{
\item{...}{slot for tablerNavMenuDropItem elements}

\item{text}{Name displayed in navbar}

\item{icon}{Item icon}
}
\description{
This items can replace a tablerNavMenuItem
}
