% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tabler.R
\name{tablerNavMenuDropItem}
\alias{tablerNavMenuDropItem}
\title{Build a tabler dashboard Dropdown item to be placed inside tablerNavMenuDrop()}
\usage{
tablerNavMenuDropItem(..., tabName = NULL)
}
\arguments{
\item{...}{Item name displayed in the UI}

\item{tabName}{Should correspond exactly to the tabName given in tablerTabItem}
}
\description{
Build a tabler dashboard Dropdown item to be placed inside tablerNavMenuDrop()
}
