% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_forestplot.R
\name{theme_forest}
\alias{theme_forest}
\title{Forestplot Theme}
\usage{
theme_forest(
  base_size = 13,
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
A custom theme used in \code{\link{forestplot}} that builts upon
\link[ggplot2]{theme_minimal}.
}
\author{
Maria Kalimeri
}
